// pop-web/src/app/shared/openapi/model/cohort.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';


export interface Cohort { 
    /**
     * Cohort population
     */
    population: number;
    /**
     * 
     */
    id: string;
    /**
     * 
     */
    createdAt: string;
    /**
     * 
     */
    updatedAt: string;
    createdBy?: User | null;
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Name of the cohort
     */
    name: string;
    includeCriteria?: object | null;
    excludeCriteria?: object | null;
    /**
     * Whether the cohort is public
     */
    isPublic?: boolean;
    /**
     * The user(s) who updated the data since its creation
     */
    updatedBy?: Array<User>;
    /**
     * Cases composing the cohort
     */
    casesIds?: Array<string>;
    /**
     * Manually added cases
     */
    manualChoicesIds?: Array<string>;
    /**
     * Frozen cases
     */
    frozenSetIds?: Array<string>;
    /**
     * Human-readable description
     */
    description: string;
}