// pop-web/src/app/shared/openapi/model/modified-resource.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ModifiedResource { 
    /**
     * Unique identifier of the modified resource
     */
    id: string;
    description?: string | null;
}